git-svn-id: https://subversion.assembla.com/svn/java_experto@102 daf995d5-a636-460e-bc68-660d2383aa96

+++ proyecto/base/tripbooker/src/tripbooker/logic/ticket/TicketServiceImp.java
@@ -77,9 +77,10 @@ public class TicketServiceImp implements ITicketService{
         boolean result = false;
         try{
             TransactionManager.getInstance().begin();
-            IUserDO userDO = DAOFactory.getInstance().getUserDAO().getUserByCode(ticketBean.getUserCode());
-            ITicketDO ticketDO = DTOMapper.getInstance().getTicketDO(ticketBean,userDO);
-            result = DAOFactory.getInstance().getTicketDAO().removeTicket(ticketDO.getTicketID());
+            ITicketDO ticketDO = DAOFactory.getInstance().getTicketDAO().getTicketByCode(ticketBean.getCode());
+            if(ticketDO != null){
+                result = DAOFactory.getInstance().getTicketDAO().removeTicket(ticketDO.getTicketID());
+            }
             TransactionManager.getInstance().commit();
         }catch(TransactionException ex){
             try{
